Do some modernizing of the code. This will hopefully make it behave better under Swift.

+++ FindPanel/OgreTextFindComponents/NSOutlineView/OgreOutlineViewAdapter.m
@@ -158,7 +158,7 @@
 - (BOOL)isHighlightable { return NO; }
 
 /* Getting structural detail */
-- (unsigned)numberOfChildrenInSelection:(BOOL)inSelection
+- (NSUInteger)numberOfChildrenInSelection:(BOOL)inSelection
 {
 #ifdef DEBUG_OGRE_FIND_PANEL
 	NSLog(@"  -numberOfChildrenInSelection: of %@", [self className]);
@@ -169,7 +169,7 @@
     return [_outlineView numberOfColumns];
 }
 
-- (id)childAtIndex:(unsigned)index inSelection:(BOOL)inSelection
+- (id)childAtIndex:(NSUInteger)index inSelection:(BOOL)inSelection
 {
 #ifdef DEBUG_OGRE_FIND_PANEL
 	NSLog(@"  -childAtIndex: of %@", [self className]);
@@ -297,7 +297,7 @@
     return [_outlineView window];
 }
 
-- (unsigned)numberOfDescendantsInSelection:(BOOL)inSelection
+- (NSUInteger)numberOfDescendantsInSelection:(BOOL)inSelection
 {
     return -1;  // indeterminate
 }
